added new auth controllers
