+++ app/Http/Controllers/Auth/ForgotPasswordController.php
@@ -1,7 +1,9 @@
 <?php
 namespace App\Http\Controllers\Auth;
+
 use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\SendsPasswordResetEmails;
+
 class ForgotPasswordController extends Controller
 {
     /*
