+++ app/Http/Controllers/Auth/RegisterController.php
@@ -1,9 +1,11 @@
 <?php
 namespace App\Http\Controllers\Auth;
+
 use App\User;
 use Validator;
 use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\RegistersUsers;
+
 class RegisterController extends Controller
 {
     /*
