+++ app/Http/Controllers/Auth/ResetPasswordController.php
@@ -1,7 +1,9 @@
 <?php
 namespace App\Http\Controllers\Auth;
+
 use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\ResetsPasswords;
+
 class ResetPasswordController extends Controller
 {
     /*
